Refs #22518 Moved qappthreadcall to its own class
This should avoid some unit test import errors, as figuremanager was being imported unnecessarily
